Added ModManagerSystem for Extracting ModManager
